+++ x/authz/keeper/msg_server.go
@@ -24,7 +24,7 @@ func (k Keeper) GrantAuthorization(goCtx context.Context, msg *types.MsgGrantAut
 
 	authorization := msg.GetGrantAuthorization()
 	// If the granted service Msg doesn't exist, we throw an error.
-	if k.router.HandlerByName(authorization.MethodName()) == nil {
+	if k.router.HandlerbyTypeURL(authorization.MethodName()) == nil {
 		return nil, sdkerrors.Wrapf(sdkerrors.ErrInvalidType, "%s doesn't exist.", authorization.MethodName())
 	}